requirements.txt
# PDF Processing
pymupdf>=1.23.0

# Image Processing
Pillow>=10.0.0

# LLM & AI
openai>=1.0.0
pydantic>=2.0.0

# UI
streamlit>=1.28.0

# Data handling
pandas>=2.0.0
openpyxl>=3.1.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
xlrd
